Make sure we write all the bytes we read
It use to be broken by errors, for example:

bytes, err := buffer.ReadString('\n')

Here, we might get EOF, but there might still be some stuff in bytes

However, when you do:

b, err := buffer.ReadByte()

if err != nil, then you can return the error straight away

+++ internals/fuzz.go
@@ -15,10 +15,10 @@ func FuzzLength(data []byte) int {
 		return 0
 	}
 	output := buf.String()
-	if len(output) != written {
+	if int64(len(output)) != written {
 		panic(fmt.Sprintf("output: %q (%d bytes), yet written = %d", output, len(output), written))
 	}
-	if written < len(data) {
+	if written < int64(len(data)) {
 		// written >= len(data) at all times (all we do is insert commas where needed)
 		panic(fmt.Sprintf("Wrote less bytes than given (written = %d) (given %q, wrote %q)", written, data, output))
 	}
